Adding the refreshing info
Hi Hadley,

Amazing book ! thank you for making it open source. I'm not sure if it's useful but I've added a quick (according to me important) note that inform the user that it's not necessary to quit (CTRL + C) the shiny app and rerun it to preview the modifications. A simple refresh does the trick. 

Best.
Fodil.